+++ comicbox/enums/comicbookinfo.py
