+++ comicbox/enums/generic/age_rating.py
