+++ comicbox/enums/generic/role.py
