+++ comicbox/enums/metroninfo.py
@@ -2,10 +2,8 @@
 
 from enum import Enum
 
-from comicbox.identifiers import (
-    ID_SOURCE_NAME_MAP,
-    IdSources,
-)
+from comicbox.enums.comicbox import IdSources
+from comicbox.identifiers import ID_SOURCE_NAME_MAP
 
 
 class MetronAgeRatingEnum(Enum):
@@ -18,32 +16,6 @@ class MetronAgeRatingEnum(Enum):
     MATURE = "Mature"
     EXPLICIT = "Explicit"
     ADULT = "Adult"
-
-
-class GenericFormatEnum(Enum):
-    """Generic Format Values."""
-
-    ANTHOLOGY = "Anthology"
-    ANNOTATION = "Annotation"
-    BOX_SET = "Box Set"
-    DIGITAL = "Digital"
-    DIRECTOR_S_CUT = "Director's Cut"
-    GIANT_SIZED = "Giant Sized"
-    GRAPHIC_NOVEL = "Graphic Novel"
-    HARD_COVER = "Hard Cover"
-    HD_UPSCALED = "HD Upscaled"
-    KING_SIZED = "King Sized"
-    MAGAZINE = "Magazine"
-    MANGA = "Manga"
-    ONE_SHOT = "One-Shot"
-    PDF_RIP = "PDF Rip"
-    PREVIEW = "Preview"
-    PROLOGUE = "Prologue"
-    SCANLATION = "Scanlation"
-    SCRIPT = "Script"
-    TRADE_PAPERBACK = "Trade Paperback"
-    WEB_COMIC = "Web Comic"
-    WEB_RIP = "Web Rip"
 
 
 class MetronFormatEnum(Enum):
