+++ comicbox/transforms/metroninfo/const.py
@@ -1,5 +1,5 @@
 """Metron Transform constants."""
 
-from comicbox.identifiers import IdSources
+from comicbox.enums.comicbox import IdSources
 
 DEFAULT_ID_SOURCE = IdSources.METRON.value
