+++ tests/schemas/test_comet.py
@@ -7,7 +7,7 @@ from types import MappingProxyType
 
 import xmltodict
 
-from comicbox.fields.enum_fields import ReadingDirectionEnum
+from comicbox.enums.comicbox import ReadingDirectionEnum
 from comicbox.formats import MetadataFormats
 from comicbox.schemas.comet import CoMetSchema
 from comicbox.schemas.comicbox import ComicboxSchemaMixin
