+++ tests/schemas/test_json.py
@@ -8,7 +8,7 @@ from types import MappingProxyType
 
 import simplejson as json
 
-from comicbox.fields.enum_fields import PageTypeEnum
+from comicbox.enums.comicinfo import ComicInfoPageTypeEnum
 from comicbox.formats import MetadataFormats
 from comicbox.schemas.comicbox import ComicboxSchemaMixin
 from comicbox.schemas.comicbox.json_schema import ComicboxJsonSchema
@@ -57,7 +57,7 @@ READ_METADATA = MappingProxyType(
             "notes": TEST_READ_NOTES,
             "page_count": 36,
             "pages": {
-                0: {"page_type": PageTypeEnum.FRONT_COVER, "size": 429985},
+                0: {"page_type": ComicInfoPageTypeEnum.FRONT_COVER, "size": 429985},
                 1: {"size": 332936},
                 2: {"size": 458657},
                 3: {"size": 450456},
@@ -139,7 +139,7 @@ READ_COMICBOX_DICT = MappingProxyType(
             },
             "pages": {
                 "00": {
-                    "page_type": PageTypeEnum.FRONT_COVER.value,
+                    "page_type": ComicInfoPageTypeEnum.FRONT_COVER.value,
                     "size": 429985,
                 },
                 "01": {"size": 332936},
