+++ tests/schemas/test_multi.py
@@ -7,7 +7,8 @@ from types import MappingProxyType
 
 import simplejson as json
 
-from comicbox.fields.enum_fields import PageTypeEnum, ReadingDirectionEnum
+from comicbox.enums.comicbox import ReadingDirectionEnum
+from comicbox.enums.comicinfo import ComicInfoPageTypeEnum
 from comicbox.formats import MetadataFormats
 from comicbox.schemas.comicbox import ComicboxSchemaMixin
 from comicbox.schemas.comicbox.json_schema import ComicboxJsonSchema
@@ -88,7 +89,7 @@ READ_METADATA = MappingProxyType(
             "bookmark": 12,
             "original_format": "Comic",
             "pages": {
-                0: {"page_type": PageTypeEnum.FRONT_COVER, "size": 429985},
+                0: {"page_type": ComicInfoPageTypeEnum.FRONT_COVER, "size": 429985},
                 1: {"size": 332936},
                 2: {"size": 458657},
                 3: {"size": 450456},
