+++ tests/test_codex_api.py
@@ -12,7 +12,7 @@ import pytest
 
 from comicbox.box import Comicbox
 from comicbox.config import get_config
-from comicbox.fields.enum_fields import PageTypeEnum
+from comicbox.enums.comicinfo import ComicInfoPageTypeEnum
 from comicbox.merge import AdditiveMerger
 from comicbox.schemas.comicbox import ComicboxSchemaMixin
 from comicbox.version import VERSION
@@ -76,7 +76,7 @@ TEMPLATE_MD = MappingProxyType(
             "notes": TEST_READ_NOTES,
             "page_count": 36,
             "pages": {
-                0: {"page_type": PageTypeEnum.FRONT_COVER, "size": 429985},
+                0: {"page_type": ComicInfoPageTypeEnum.FRONT_COVER, "size": 429985},
                 1: {"size": 332936},
                 2: {"size": 458657},
                 3: {"size": 450456},
